+++ commands/barter.mjs
@@ -1,9 +1,6 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
-import {
-    MessageEmbed,
-} from 'discord.js';
+import { MessageEmbed } from 'discord.js';
 
-import getCurrencies from '../modules/get-currencies.mjs';
 import getCraftsBarters from '../modules/get-crafts-barters.mjs';
 import progress from '../modules/progress.mjs';
 
@@ -36,7 +33,6 @@ const defaultFunction = {
         const matchedBarters = [];
 
         const { barters } = await getCraftsBarters();
-        const currencies = getCurrencies();
 
         for (const barter of barters) {
             if (barter.rewardItems[0].item.name.toLowerCase().includes(searchString.toLowerCase())) {
@@ -97,7 +93,7 @@ const defaultFunction = {
                     if (!offer.vendor.trader) {
                         continue;
                     }
-                    let traderPrice = offer.price * currencies[offer.currency];
+                    let traderPrice = offer.priceRUB;
 
                     if ((traderPrice < itemCost && prog.traders[offer.vendor.trader.id] >= offer.vendor.minTraderLevel) || itemCost == 0) {
                         itemCost = traderPrice;
