+++ commands/map.mjs
@@ -1,9 +1,8 @@
-import {
-    SlashCommandBuilder
-} from '@discordjs/builders';
+import { SlashCommandBuilder } from '@discordjs/builders';
+import { MessageEmbed } from 'discord.js';
 
 import gameData from '../modules/game-data.mjs';
-import getMapEmbed from '../modules/get-map-embed.js';
+import realTimeToTarkovTime from '../modules/time.mjs';
 
 const defaultFunction = {
     data: new SlashCommandBuilder()
@@ -20,7 +19,53 @@ const defaultFunction = {
         await interaction.deferReply();
         const mapId = interaction.options.getString('map');
 
-        const embed = await getMapEmbed(mapId);
+        const mapData = await gameData.maps.getAll();
+        const embed = new MessageEmbed();
+
+        const selectedMapData = mapData.find(mapObject => mapObject.id === mapId);
+        let displayDuration = `${selectedMapData.raidDuration} minutes`;
+
+        // Get left and right real tarkov time
+        let left = realTimeToTarkovTime(new Date(), true);
+        let right = realTimeToTarkovTime(new Date(), false);
+        let displayTime = `${left} - ${right}`;
+        if (selectedMapData.name.includes('Factory')) {
+            // If the map is Factory, set the times to static values
+            if (selectedMapData.name.includes('Night')) {
+                displayTime = '03:00';
+            } else {
+                displayTime = '15:00';
+            }
+        } 
+
+        let displayPlayers = '???';
+        if (selectedMapData.players) {
+            displayPlayers = selectedMapData.players;
+        }
+
+        let mapUrl = false; `https://tarkov.dev/map/${selectedMapData.key}`;
+        if (selectedMapData.key) {
+            mapUrl = `https://tarkov.dev/map/${selectedMapData.key}`;
+        } else if (selectedMapData.wiki) {
+            mapUrl = selectedMapData.wiki;
+        }
+
+        // Construct the embed
+        embed.setTitle(selectedMapData.name);
+        if (mapUrl) {
+            embed.setURL(mapUrl);
+        }
+        embed.addField('Duration ⌛', displayDuration, true);
+        embed.addField('Players 👥', displayPlayers, true);
+        embed.addField('Time 🕑', displayTime, true);
+        if (selectedMapData.key) {
+            embed.setImage(`https://tarkov.dev/maps/${selectedMapData.key}.jpg`);
+        }
+
+        // If the map was made by a contributor, give them credit
+        if (selectedMapData.source) {
+            embed.setFooter({ text: `Map made by ${selectedMapData.source}` });
+        }
 
         await interaction.editReply({
             embeds: [embed],
