--- modules/all-maps.js
@@ -1,13 +1,0 @@
-const allMaps = [
-    ['customs', 'customs'],
-    ['factory', 'factory'],
-    ['factory (night)', 'factory (night)'],
-    ['interchange', 'interchange'],
-    ['labs', 'labs'],
-    ['lighthouse', 'lighthouse'],
-    ['reserve', 'reserve'],
-    ['shoreline', 'shoreline'],
-    ['woods', 'woods']
-];
-
-export default allMaps;
